describe dispute game k-section mechanism
This describes an alternative mechanism to Arbitrum's bisection
mechanism, which would have the sequencer commit to the root of a
binary Merkle tree encoding the entire execution trace of some
computation.

Instead, the sequencer simply commits to equally spaced machine steps
spanning a shrinking interval.

This seems to be a simpler and more economical way to quickly zoom into
the first contested machine state transition, demonstrated by the
prototype Typescript implementation. (I chose Typescript, as that's
the most recent language I've used for development, and I wanted
to include running tests.)