opnode: Update sync start algorithm
The sync start algorithm has been modified to find both a SafeL2Head
and UnsafeL2Head. The UnsafeL2Head is the first L2 block that has
an L1 origin that is canonical in the new chain.

The FindL2Heads function first finds the UnsafeL2Head, and then continues
to walk back the L2 chain until it finds and L2 block that can be completely
derived from the sequence window that has not been changed in the
reorg (i.e. the last block of the window was the reorg base).

This FindSyncStart function previously only found the UnsafeHead.