opnode: Allow L2 chain to run ahead of L1
This allows the sync algorithm to return the L2 start block that
has an L1 Origin that is ahead of the L1 Head.

Note that we still base the safe head as 1 sequencing window back
from the first L2 block that is fully canonical. This is still
required to make sure that we don't miss a change to the sequencing
window.