bugfix: L1StandardBridge finalizeETHWithdrawal payable
The `L1StandardBridge` does not hold on to ETH balances,
the `OptimismPortal` does so. When depositing/withdrawing
ETH through the standard bridge, the `OptimismPortal` must
forward ETH to the `L1StandardBridge` for it to be forwarded
to the user. This means that the `finalizeETHWithdrawal`
function needs to be `payable`. This commit pulls in the
interfaces for the `IL1ERC20Bridge` and the `IL1StandardBridge`
contracts and adds a `payable` modifier to `finalizeETHWithdrawal`.
This is a diff from the previous deployment of the contracts,
but it does not impact the ABI so there is no diff for the
end user because this function is only callable by a smart contract.